refactor(types): extract StandardAttributesResolver — MCE coupling 24→19
