+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/types/registry/MetadataCollectionSpecializer.java
@@ -196,8 +196,8 @@ public class MetadataCollectionSpecializer {
       MetadataChildrenExtractor::attributesFor),
     new CollectionSpec("СтандартныеРеквизиты", "StandardAttributes",
       BASE_COLLECTION_STD_ATTR, "ОписаниеСтандартногоРеквизита",
-      MetadataChildrenExtractor::hasStandardAttributes,
-      MetadataChildrenExtractor::standardAttributesFor),
+      StandardAttributesResolver::hasStandardAttributes,
+      StandardAttributesResolver::standardAttributesFor),
     new CollectionSpec("ТабличныеЧасти", "TabularSections",
       BASE_COLLECTION_METADATA, "ОбъектМетаданных: ТабличнаяЧасть",
       MetadataMdoPredicates::isTabularSectionOwner,
